Rename: 로컬 컴포넌트 파일명 및 폴더 구조 변경, 경로 수정
