--- src/app/assignment/(components)/AssignmentButton.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const AssignmentButton = () => {
-  return <div>AssignmentButton</div>;
-};
-
-export default AssignmentButton;
