+++ src/app/assignment/(components)/AssignmentStudentViewCard.tsx
@@ -1,12 +1,12 @@
 "use client";
 
 import React, { useState } from "react";
-import ProfileImage from "./ProfileImage";
-import Modal from "./Modal";
-import SubmitAssignmentWithLink from "./SubmitAssignmentWithLink";
-import SubmitAssignmentWithFile from "./SubmitAssignmentWithFile";
+import AssignmentProfileImage from "./AssignmentProfileImage";
+import AssignmentModal from "./AssignmentModal";
+import AssignmentSubmitWithLink from "./AssignmentSubmitWithLink";
+import AssignmentSubmitWithFile from "./AssignmentSubmitWithFile";
 
-const StudentViewCard: React.FC = () => {
+const AssignmentStudentViewCard: React.FC = () => {
   const [isLinkOpen, setIsLinkOpen] = useState(false);
   const [isFileOpen, setIsFileOpen] = useState(false);
 
@@ -14,7 +14,7 @@ const StudentViewCard: React.FC = () => {
     <>
       <div className="flex justify-between items-center px-[21px] py-[24px] border rounded-[10px]">
         <div className="flex justify-start items-center gap-[14px]">
-          <ProfileImage />
+          <AssignmentProfileImage />
           <div>
             <div className="mb-[5px]">
               <span className="mr-[15px] text-grayscale-100 text-[16px] font-[700]">
@@ -63,36 +63,36 @@ const StudentViewCard: React.FC = () => {
         </div>
       </div>
       {/* 과제: 링크 제출 */}
-      <Modal
+      <AssignmentModal
         title="과제 제출"
         isOpen={isLinkOpen}
         onClose={() => {
           setIsLinkOpen(false);
         }}
       >
-        <SubmitAssignmentWithLink
+        <AssignmentSubmitWithLink
           onClose={() => {
             setIsLinkOpen(false);
           }}
         />
-      </Modal>
+      </AssignmentModal>
 
       {/* 과제: 파일 제출 */}
-      <Modal
+      <AssignmentModal
         title="과제 제출"
         isOpen={isFileOpen}
         onClose={() => {
           setIsFileOpen(false);
         }}
       >
-        <SubmitAssignmentWithFile
+        <AssignmentSubmitWithFile
           onClose={() => {
             setIsFileOpen(false);
           }}
         />
-      </Modal>
+      </AssignmentModal>
     </>
   );
 };
 
-export default StudentViewCard;
+export default AssignmentStudentViewCard;
