--- src/app/assignment/list/page.tsx
@@ -1,14 +1,0 @@
-"use client";
-
-import React from "react";
-import ListContent from "../components/ListContent";
-
-const List = () => {
-  return (
-    <div>
-      <ListContent />
-    </div>
-  );
-};
-
-export default List;
